+++ includes/classes/AdminNotices.php
@@ -418,6 +418,10 @@ class AdminNotices {
 	 * @return array|bool
 	 */
 	protected function process_es_below_compat_notice() {
+		if ( Utils\is_epio() ) {
+			return false;
+		}
+
 		$host = Utils\get_host();
 
 		if ( empty( $host ) ) {
@@ -474,6 +478,10 @@ class AdminNotices {
 	 * @return array|bool
 	 */
 	protected function process_es_above_compat_notice() {
+		if ( Utils\is_epio() ) {
+			return false;
+		}
+
 		$host = Utils\get_host();
 
 		if ( empty( $host ) ) {
@@ -525,6 +533,12 @@ class AdminNotices {
 			return false;
 		}
 
+		$screen = Screen::factory()->get_current_screen();
+
+		if ( 'install' === $screen ) {
+			return false;
+		}
+
 		$es_version = Elasticsearch::factory()->get_elasticsearch_version( false );
 
 		if ( false !== $es_version ) {
@@ -532,7 +546,7 @@ class AdminNotices {
 		}
 
 		// Only dismissable on non-EP screens
-		if ( ! in_array( Screen::factory()->get_current_screen(), [ 'settings', 'dashboard' ], true ) ) {
+		if ( ! in_array( $screen, [ 'settings', 'dashboard' ], true ) ) {
 			if ( defined( 'EP_IS_NETWORK' ) && EP_IS_NETWORK ) {
 				$dismiss = get_site_option( 'ep_hide_host_error_notice', false );
 			} else {
@@ -591,6 +605,16 @@ class AdminNotices {
 		}
 
 		if ( defined( 'EP_IS_NETWORK' ) && EP_IS_NETWORK ) {
+			$last_sync = get_site_option( 'ep_last_sync', false );
+		} else {
+			$last_sync = get_option( 'ep_last_sync', false );
+		}
+
+		if ( empty( $last_sync ) ) {
+			return false;
+		}
+
+		if ( defined( 'EP_IS_NETWORK' ) && EP_IS_NETWORK ) {
 			$dismiss = get_site_option( 'ep_hide_yellow_health_notice', false );
 		} else {
 			$dismiss = get_option( 'ep_hide_yellow_health_notice', false );
@@ -604,7 +628,7 @@ class AdminNotices {
 
 		$nodes = Stats::factory()->get_nodes();
 
-		if ( $nodes < 2 ) {
+		if ( false !== $nodes && $nodes < 2 && $nodes > 0 ) {
 			if ( defined( 'EP_IS_NETWORK' ) && EP_IS_NETWORK ) {
 				$url = network_admin_url( 'admin.php?page=elasticpress-health' );
 			} else {