+++ includes/classes/Feature/SearchOrdering/SearchOrdering.php
@@ -740,6 +740,10 @@ class SearchOrdering extends Feature {
 		$pointers = get_post_meta( $post_id, 'pointers', true );
 		$term     = $this->create_or_return_custom_result_term( $post->post_title );
 
+		if ( empty( $pointers ) ) {
+			return;
+		}
+
 		foreach ( $pointers as $pointer ) {
 			$ref_id = $pointer['ID'];
 			wp_remove_object_terms( $ref_id, (int) $term->term_id, self::TAXONOMY_NAME );